uchar/c32rtomb: Explicitly handle s = NULL in c32rtomb() function
c32rtomb() and wcsnrtombs() have different behaviors when the destination buffer
(s or dst) is NULL. This difference can lead to issues, especially when dealing
with surrogate pairs or characters that require multiple bytes in their
multibyte representation.

c32rtomb():
According to the C standard C11 7.28.1 when s is NULL, the c32rtomb function is
equivalent to the call c32rtomb(buf, L'\0', ps) that resets the state
(mbstate_t) to the initial state and returns 1
(same for wcrtomb() function). This is because c32rtomb() and wcrtomb()
are designed to handle a single char32_t character at a time,
and when s is NULL, it doesn't perform any conversion but simply resets
the state.

wcsnrtombs():
when dst is NULL, wcsnrtombs() calculates the total number of bytes
required to store the entire wide character string as a multibyte sring.
it does not reset the state or return 1. instead, it processes the entire string
and returns the total byte count.

Signed-off-by: Ahmed Shehab <shehab@esynopsys.com>