zephyr: cmake: Use correct Kconfig for Zephyr's -Oz abstraction
An abstraction for -Oz was added to Zephyr in [1] and a mapping
for picolibc was added in [2]. Part way through [1], the Kconfig
value was renamed from CONFIG_MIN_SIZE_OPTIMIZATIONS to
CONFIG_SIZE_OPTIMIZATIONS_AGGRESSIVE. [2] was changed accordingly,
but it looks like this was accidentally droppped/reverted in the
final rebase before [2] was merged.

Currently, Zephyr's hello_world samples are failing to build for me
where CONFIG_MIN_SIZE_OPTIMIZATIONS and CONFIG_PICOLIBC_USE_MODULE
are specified with undefined references to memcpy due to this. So,
make sure we are using the correct Kconfig value.

[1] https://github.com/zephyrproject-rtos/zephyr/pull/70511
[2] https://github.com/picolibc/picolibc/pull/737

Signed-off-by: Jonathon Penix <jpenix@quicinc.com>