Silence -Wshift-count-overflow warnings
This patch fixes a few "left shift count >= width of type
[-Wshift-count-overflow]" warnings. Before shifting a char 16 (or more)
bits left first it explicitly casts the char to uint32_t. The existing
code relies on implicit casts to int and assumes that ints are 32-bit.
This is not always true because the C standard does not require int to
be 32-bit and there are targets (e.g. H8/300) where by default int is
indeed 16-bit.

2025-03-02  Jan Dubiec  <jdx@o2.pl>

newlib/ChangeLog:

	* libc/stdlib/gdtoa-gdtoa.c (gdtoa): Cast to __ULong before left shift.
	* libc/string/memmem.c (memmem): Cast to uint32_t before left shift.
	* libc/string/strstr.c (strstr2): Ditto.
	(strstr3): Ditto.
	(strstr4): Ditto.

 newlib/libc/stdlib/gdtoa-gdtoa.c | 2 +-
 newlib/libc/string/memmem.c      | 3 ++-
 newlib/libc/string/strstr.c      | 6 +++---
 3 files changed, 6 insertions(+), 5 deletions(-)