picolib/powerpc: Add _set_tls
Like m68k, powerpc sets the TLS base 0x7000 above the TLS storage so
that a signed 16-bit offset can reach a large amount of space. A TCB
could be placed below that. Powerpc also reserves register 13 for this
value.

The stack protector places the check value *below* the TLS block;
before TLS was supported, this was handled by setting the TLS base
register (r13) to a value suitable for reaching the __stack_chk_guard
variable. Now, this value becomes just another member of that
block.

To make that work, we allocate space at the end of the .data section
and then add the size of that to various computed values so that the
TLS block appears to be larger to the picolibc code but not to the linker
so that it assigns correct TLS offsets.

Signed-off-by: Keith Packard <keithp@keithp.com>