Cygwin: stdlib.h: resurrect arc4random_stir() definition
An autoconf test for arc4random_stir() succeeds on Cygwin, because
we export the symbol for backward compatibility.  However, when
building with -Werror afterwards, the definition of arc4random_stir()
is missing in the stdlib.h header, so there's a chance to encounter
an error like this with newer GCC:

sshd.c:1113:25: error: implicit declaration of function ‘arc4random_stir’; did you mean ‘arc4random_buf’? [-Wimplicit-function-declaration]
 1113 |                         arc4random_stir();
      |                         ^~~~~~~~~~~~~~~
      |                         arc4random_buf

This occurs for instance when building OpenSSH 9.9p2 with GCC 15.0.

Define arc4random_stir() as an empty macro in cygwin/stdlib.h
to avoid the above compiler warning.

Do not define arc4random_addrandom() yet, in the hope that this function
call is really, really not used anymore.

Fixes: 1cca343e4714 ("Drop redundant arc4random prototypes from cygwin/stdlib.h")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>