Cygwin: pipe: Fix 'lost connection' issue in scp
When transferring huge file using scp, the "lost connection" error
sometimes happen. This is due to fhandler_pipe_fifo::raw_write()
accidentally sends data that is not reported in th return value when
interrupted by a signal. The cause of the problem is that CancelIo()
responds success even if NtWriteFile() already sends the data.

The following testcase using plain Win32 APIs reproduces the issue.
The output will be something like:
W: 8589934592
R: 9280061440
Much more data was received than the sender thought it had sent.

#include <windows.h>
#include <stdio.h>

DWORD WINAPI read_thread(LPVOID arg)
{
  char buf[32768];
  ssize_t sum = 0;
  HANDLE pr = (HANDLE) arg;
  while (1) {
    DWORD nb;
    BOOL res = ReadFile(pr, buf, sizeof(buf), &nb, NULL);
    if (!res) break;
    sum += nb;
  }
  printf("R: %lld\n", sum);
}

int main()
{
  const char pipename[] = "\\\\.\\pipe\\testpipe";
  char buf[65536*4] = {0,};
  OVERLAPPED ov = {0,};
  ssize_t sum = 0;

  HANDLE pw = CreateNamedPipe(pipename,
                PIPE_ACCESS_OUTBOUND | FILE_FLAG_OVERLAPPED,
                PIPE_TYPE_BYTE, 1, 65536, 0, 0, NULL);
  HANDLE pr = CreateFile(pipename, GENERIC_READ, 0, NULL, OPEN_EXISTING,
                FILE_ATTRIBUTE_NORMAL, NULL);
  HANDLE thr = CreateThread(NULL, 0, read_thread, pr, 0, NULL);
  ov.hEvent = CreateEvent(NULL, FALSE, FALSE, NULL);

  int cnt = 0;
  while (sum < 8192LL*1024*1024) {
    DWORD nb;
    BOOL res = WriteFile(pw, buf, sizeof(buf), &nb, &ov);
    if (res) sum += nb;
    else if (GetLastError() == ERROR_IO_PENDING) {
      if ((cnt & 3) == 0) CancelIo (pw);
      res = GetOverlappedResult(pw, &ov, &nb, TRUE);
      if (!res && GetLastError() != ERROR_OPERATION_ABORTED) {
        printf("%08x\n", GetLastError());
        break;
      }
      sum += nb;
    }
    cnt++;
  }

  printf("W: %lld\n", sum);
  CloseHandle(pw);
  CloseHandle(ov.hEvenet);

  WaitForSingleObject(thr, INFINITE);
  CloseHandle(pr);

  return 0;
}

It seems that CancelIo() works as expected when the NtWriteFile()
is stucked on the pipe full. If the writing is on-going, CancelIo()
reports cancelled, but the data may already be written.

To avoid this issue, this patch waits for completion of writing a
while before issuing CancelIo(), and if it has not completed yet,
call CancelIo().
The wait is done using existing cygwait(select_sem, 10, cw_cancel).
If the semaphore select_sem is released in the read side, the write
operation will be on-going. Otherwise, the write operation is still
stucking, so CancelIo() can safely cancel the I/O.

Addresses: https://cygwin.com/pipermail/cygwin/2025-January/257143.html
Fixes: 4003e3dfa1b9 ("Cygwin: pipes: always terminate async IO in blocking mode")
Reported-by: Jay M Martin <jaymmartin_buy@cox.net>
Reviewed-by: Corinna Vinschen <corinna@vinschen.de>
Signed-off-by: Takashi Yano <takashi.yano@nifty.ne.jp>