Cygwin: signals: don't evaluate SA_SIGINFO context after handler returns
After return from a signal handler created with the SA_SIGINFO flag,
Cygwin checks if the handler changed the context.  If so, Cygwin
erroneously uses the context in a call to setcontext().

Drop this special behaviour.  Neither Linux man pages, nor POSIX
documentation treat the context pointer given to the handler as
anything other than a read-only area.

Signed-off-by: Corinna Vinschen <corinna@vinschen.de>