Cygwin: sigdelayed: pop return address from signal stack earlier
When a signal handler is called via sigdelayed, sigdelayed pops
the original address to return to from the signal stack.  It only
does so after the signal handler has been called and returned.

If the signal handler never returns, for instance because it calls
longjmp or setcontext, the address stays on the signal stack and
the stack loses one slot.  Given the stack has a depth of 256 entries,
no signal handler will be called after 256 signal handlers long
jumping out.

Pop the return address from the signal stack prior to calling the
signal handler and store it in a callee-saved register.  This avoids
filling up the signal stack in a scenario like this.

Addresses: https://cygwin.com/pipermail/cygwin/2025-March/257648.html
Fixes: 61522196c715 ("* Merge in cygwin-64bit-branch.")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>