Cygwin: stdlib.h: revert declaring arc4random_stir() as function
The idea to define arc4random_stir() as macro to avoid autoconf
problems in commit a8891c932192 ("Cygwin: stdlib.h: resurrect
arc4random_stir() definition") worked fine for OpenSSH, but not
for libbsd.  During the build process the macro shadows the
definition of arc4random_stir() and results in an error message
like this:

  error: macro "arc4random_stir" passed 1 arguments, but takes just 0

Give up and just declare arc4random_stir() in a compatible way with
libbsd again.

Fixes: a8891c932192 ("Cygwin: stdlib.h: resurrect arc4random_stir() definition")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>