semihost/x86: add TLS support for semihosting Intel i386
On the i386, the precise procedure for setting the current
TLS block pointer -- the %gs segment base -- is
unfortunately OS-specific, since it generally involves
privileged operations.  The proposed simplistic
implementation is meant to work in the bare metal
environment set up by picocrt/x86.

The linker script was also modified to allocate space for
the thread control block, which on the i386 appears _after_
the TLS variables (https://akkadia.org/drepper/tls.pdf).

Signed-off-by: tkchia <tkchia@noreply.codeberg.org>