machine/arc: strncpy fails as __strncpy_bzero passes target addr in r3, not r0
The memset code has a special entry point used to clear space beyond
the destination in strncpy. To make it work, the destination is passed
in r3 instead of r0 (as r0 must hold the final return value).

However, the code to clear the unaligned end of the string was
mistakenly computing the end address using r0 instead of r3. If the
number of bytes to erase was smaller than the copied string, that
would end up smashing bytes in the middle of the result instead of
clearing bytes at the end.

Signed-off-by: Keith Packard <keithp@keithp.com>