math: Clean up float predicates
Each predicate (isinf, isnan, etc) needs to exist in binary form for
each of the three types (float, double, long double) and each needs to
have two names, one with two leading underscores (and a trailing 'd'
for the double version) and another without leading underscores.

We can't usually use the compiler builtins because they raise INVALID
conditions when passed a sNAN, so instead there are custom versions for
each type.

Finally, there need to be macros for each which work with all
types. These could be written using the C _Generic mechanism, but that
relies on a relatively recent C compiler. Instead, we just check the
size of the operand and assume that will uniquely identify the type.

Signed-off-by: Keith Packard <keithp@keithp.com>