machine/arc: non-bs strncpy updates dest addr inconsistently
Each half of the 8-byte loop iteration needs to leave the destination
address pointing after the last 4 written bytes, instead of having the
first half move 8 bytes and the second half move zero bytes. With this
fix, the byte-at-a-time loop at the end will start with the correct
address even when the 8-byte loop terminates in the first half.

Signed-off-by: Keith Packard <keithp@keithp.com>