picocrt/powerpc: Initialize iplt table at startup
This table contains copies of the function descriptors that are
accessed via the TOC instead of directly from memory. For a static
executable, the contents are static and could be computed at link
time. However, the linker doesn't do that so instead we need to
run a short loop at program startup time.

Because the iplt entries must be reachable from the tocbase register
using 16-bit offsets, they need to be close to the rest of the TOC,
which we (currently) stick in read-only memory. As a kludge, stick the
iplt fields there too. This "works" because powerpc targets don't have
read-only memory.

Signed-off-by: Keith Packard <keithp@keithp.com>