tinystdio: Finish switchover to _PICOLIBC_PRINTF and _PICOLIBC_SCANF
These two values replace the list of per-variant
PICOLIBC_*_PRINTF_SCANF variables. They provide a simpler mechanism
for the compiler driver to pass along the right value and select the
right linker aliases.

Switch the code to use the new values instead of the old ones,
providing backward-compatibility hooks in stdio.h so that existing
external code continues to work.

Signed-off-by: Keith Packard <keithp@keithp.com>