locale: drop global_locale_string, add locale_string to locale_t
After 71511d4ac868 ("getlocalename_l: implement per SUS Base
Specifications Issue 8 draft") the Austin Group changed the definition
of getlocalename_l() to return setlocale-compatible strings if LC_ALL
has been specified.  In contrast to the former definition which
disallowed LC_ALL as category.

In preparation, add a locale_string to every locale_t, and drop
the static global_locale_string.

This in turn requires to change currentlocale() to work locale-agnostic.
Rename to  __currentlocale and take two parameters, the locale object
and a target string to store the locale string.  The latter is required
to allow specifiying a target string not in the current locale.

Link: https://www.austingroupbugs.net/view.php?id=1741
Fixes: 71511d4ac868 ("getlocalename_l: implement per SUS Base Specifications Issue 8 draft")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>