Cygwin: exceptions: check ExceptionFlags for EXCEPTION_NONCONTINUABLE
So far, exception::handle returned prematurely if the value of
EXCEPTION_RECORD::ExceptionFlags was non-0.

Starting with Windows 11 we're getting into trouble with that,
if the exception is software generated, for instance by calling
"throw" in a C++ application.

In that case EXCEPTION_SOFTWARE_ORIGINATE (0x80) is set in
EXCEPTION_RECORD::ExceptionFlags.

Change the condition for exiting prematurely to do this only if
any other flag except EXCEPTION_SOFTWARE_ORIGINATE is set in
EXCEPTION_RECORD::ExceptionFlags.

Fixes: Silent change in Windows exception handling
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>