Cygwin: dlfcn: add native DLLs to dll list
If native DLLs are dlopen'ed, we have to keep track of the load
count, so dlclose doesn't misbehave.  We also have to keep track
for the sake of a forked child so the handle returned by dlopen
is still valid in the child.

To accomplish that, add a new type DLL_NATIVE.

Fixes: 82b31085f6ad ("Cygwin: dlfcn: avoid ENOENT on dlcose after dlopen(cygwin1.dll)")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>