tinystdio: Implement wcsto* string to number functions
Directly implement all of the wide char numeric conversion functions
by extending the existing functions to also support being built with
wide chars, just like the printf and scanf code do.

These replace the stdlib versions, some of which allocated a temporary
buffer to use with wcstombs along with making the implementations all
stand-alone.

Test wcstol, wcstof and wcstod to make sure the basic structure of the
changes are correct.

Signed-off-by: Keith Packard <keithp@keithp.com>