cdefs: Create separate __deprecated_m that takes a message
GCC supports two forms of the deprecated attribute, one with no
argument that prints a generic message and another with a string
argument that is displayed as part of the warning. Zephyr uses only
the argument-free version and calls that __deprecated.

To avoid conflicting with that, replicate that behavior in picolibc
and then create another macro, __deprecated_m(s), which takes the
message and can be used in the one place in picolibc headers which
expose a deprecated function (mktemp).

Signed-off-by: Keith Packard <keithp@keithp.com>