ubsan: Provide complete set of handlers for -fsanitize=undefined
These are all of the undefind behavior handlers for both gcc and
clang. When building applications with -fsanitize=undefined, these
will present error messages and call abort() to terminate the
application.

Negative shift operations with valid values are allowed so that
arithmetic right shift continues to work. Right shift is
implementation defined, while left shift is undefined behavior, but
there doesn't appear to be a way to distinguish these in the handler.

This uses the existing meson -Db_sanitize option while leaving the
-Dsanitize=bounds option available as a compatibility change.

Signed-off-by: Keith Packard <keithp@keithp.com>

fixup-flags