math: Add compliant implementations of shifts for signed integers
Left shift of negative integers is undefined behavior (a very
programmer-hostile part of the C standard). Work around that by
creating a macro which performs the shift by casting to the related
unsigned type, performing the shift, and then casting back.

Right shift of negative integers is implementation defined. Known
compilers all implement this as arithmetic shift, which is what we
want. To make the sanitizer happy, wrap this in a macro which does the
computation using unsigned shifts when the sanitizer is enabled and
otherwise just uses the shift operator and expects to get arithmetic
shifts.

Signed-off-by: Keith Packard <keithp@keithp.com>