tinystdio: Disable sanitizer in asprintf code paths
Asprintf needs to compute the address of the start of the returned
buffer; the FILE contains only the size of the buffer and a pointer to
the end. Subtracting the size from the end produces the correct
result, but the address sanitizer doesn't believe it. Disable the
sanitizer entirely for these functions.

Signed-off-by: Keith Packard <keithp@keithp.com>