semihost/x86: add TLS support for semihosting Intel i386 and x86-64
On the x86, the precise procedure for setting the current
TCB pointer -- the %gs segment base (for i386) or the
%fs base (for x86-64) -- is unfortunately OS-specific, since
it generally involves privileged operations.

The current proposed arrangement is to have an OS-indepedent
`_set_tls()` routine calculate the TCB address from the TLS
base -- by adding the needed offset -- and then simply hand
over to a lower layer `__set_tcb()` routine to actually set
the TCB pointer.

(The `__set_tcb()` interface follows that of OpenBSD: see
https://man.openbsd.org/__get_tcb.2 .)

The linker script was also modified to allocate space for
the thread control block, which on the x86 appears _after_
the TLS variables (https://akkadia.org/drepper/tls.pdf).

Signed-off-by: tkchia <tkchia@noreply.codeberg.org>