tinystdio: ryu fixup table for (1/i)^5 is too small and missing values
This table came from the original ryu code; it seems like the C
translation of the Java code didn't adapt the extended range used in C
to the compact implementation, so it was missing fixup entries for
values > 288.

Those were easily computed using the full table and comparing the
results to the compact table and inserting the necessary correction
factors.

Signed-off-by: Keith Packard <keithp@keithp.com>