meson: Rework how compiler link tests are done
When performing meson compiler tests that use linking, we need to pass
along an alias for '_start' so that linking doesn't rely up on any
startup code. Share the code which detects which command line form
aliases require to do this.

Stick this argument, along with any runtime library (libgcc or
libclang_rt.builtins) in the 'link_test_args' variable for use
by other tests.

Signed-off-by: Keith Packard <keithp@keithp.com>