meson: Don't install picolibc if picolib is a subproject
Currently if picolibc is used as a subproject triggering install
on the outter project will cause all of the targets that picolibc
provides that are installable will also get installed.

This is probably undesirable as it'll replace whatever picolibc
the user has already installed, the picolibc configured in the
subproject is probably explictly for the use of the outter project,
picolibc is probably statically linked into the output of the
outter project so there is no need to install anything from
picolibc etc etc etc.

Unfortunately, meson doesn't have a good way of disabling
installation of subproject targets from the outter project
so we need to check if picolibc is a subproject and remove
the install flag/not define the target for everything.

Signed-off-by: Daniel Palmer <daniel@0x0f.com>