Make Armv7-A memcpy work in Thumb state
The implementation of Armv7-A's `memcpy` requires Arm state, therefore
callers from Thumb state must perform an interworking call by the use of
`BLX`.

Before this patch, the alias declaration for `__aeabi_memcpy(4|8)?` used
a macro (`ASM_ALIAS`) that, when targeting Thumb, used the directive
`.thumb_set`.  This directive, in this specific case, wrongly marked the
alias as a Thumb entry point.

This patch defines a new macro to define an alias unconditionally for
the Arm state, and uses this new macro to define the aliases to
`memcpy`.