Fixes to make AArch32 A-profile work better in Thumb state
 - Make Armv7-A memcpy work in Thumb state

The implementation of Armv7-A's `memcpy` requires Arm state, therefore
callers from Thumb state must perform an interworking call by the use of
`BLX`.

Before this patch, the alias declaration for `__aeabi_memcpy(4|8)?` used
a macro (`ASM_ALIAS`) that, when targeting Thumb, used the directive
`.thumb_set`.  This directive, in this specific case, wrongly marked the
alias as a Thumb entry point.

This patch defines a new macro to define an alias unconditionally for
the Arm state, and uses this new macro to define the aliases to
`memcpy`.

 - AArch32 A-profile using Thumb state must stay at it on exception entry

On exception entry, the architecture may switch ISA depending on the TE
bit of the SCTLR register.

Before this patch, the startup code was overprescriptive and set the TE
bit only if the target is Thumb but not A-profile. This last condition
isn't correct: A-profile architectures may run on Thumb state too.

This change fixes the preprocessor conditional to test only
`__thumb2__`.

Moreover, the second read of SCTLR needs to be marked as volatile in
inline assembly to prevent the compiler from optimizing it away. Without
this mark, the compiler presumes that the second read is redundant
because it isn't aware of the write that happens in between.

 - AArch32 A-profile interrupt vectors in Thumb2 mode should use Thumb2 instructions

This code assumes that A-profile can't run in Thumb mode. As a
consequence, code that does run in it faults if exceptions are taken.

This patch changes the conditional to check only for Thumb mode.