AArch32 A-profile using Thumb state must stay at it on exception entry
On exception entry, the architecture may switch ISA depending on the TE
bit of the SCTLR register.

Before this patch, the startup code was overprescriptive and set the TE
bit only if the target is Thumb but not A-profile. This last condition
isn't correct: A-profile architectures may run on Thumb state too.

This change fixes the preprocessor conditional to test only
`__thumb2__`.

Moreover, the second read of SCTLR needs to be marked as volatile in
inline assembly to prevent the compiler from optimizing it away. Without
this mark, the compiler presumes that the second read is redundant
because it isn't aware of the write that happens in between.