scripts: test riscv 'large' model addressing range
Place data well outside of the 20-bit medany range to validate
large model addressing. Increase memory size in qemu to cover the
additional range. We don't know if the executable is using large mode,
so just make it large enough all of the time.

Leave space above the application for the DTB.

Signed-off-by: Keith Packard <keithp@keithp.com>