picocrt: Add infrastructure for using separate section for initial TLS block
This allows the TLS block to be allocated separately from .data and .bss, at
rather than requiring that it be placed between the .data and .bss sections in
the linker script. This permits the linker script to be much simpler,
easing compatibility with existing scripts.

For this to be used, the linker script must be adapted to perform the
TLS allocation without allocating RAM, and place the .tdata
initialization data in flash. Then the startup code must call
_init_tls.

Signed-off-by: Keith Packard <keithp@keithp.com>