picocrt: Add infrastructure for using stack for initial TLS block
This allows the TLS block to be allocated out of the stack area at
startup time, rather than requiring that it be statically allocated in
the linker script. This permits the linker script to be much simpler,
easing compatibility with existing scripts.

For this to be used, each architecture must perform the TLS allocation
by using the ALLOC_TLS macro to adjust the stack pointer and passing
the stack pointer value to the _cstart function which calls _init_tls
to initialize the contents and _set_tls to set the TLS base address.

Signed-off-by: Keith Packard <keithp@keithp.com>