time: Make strftime match glibc
The big difference here is century numbers; we need to make
years from -1 to -100 use century -1 instead of century 0, and then
adjust them to be offset from year -100 instead of year 0. That makes
%C * 100 + %y always equal %Y. %C also has no minimum width.

Recurse for %D to make sure the resulting %y component is computed
correctly.

%F is really just %Y-%m-%d, the year can have fewer than 4 digits.

Move (and fix) the tests to test/test-strftime.c

Signed-off-by: Keith Packard <keithp@keithp.com>