Replace meson built-in 'b_sanitize' with 'sanitize'
Meson 1.9.0 now validates the selected sanitizer option with both the
compiler and linker, but the linker check requires a functioning C
library install to work, causing it to fail when building the C
library.

Replace that with our own 'sanitize' flag; we already checked to make
sure the compiler supported it, so all we've done now is skip the
linker check.

Signed-off-by: Keith Packard <keithp@keithp.com>