machine/riscv: Refactor machine specific memmove overriding mechanism
For RISC-V, it is possible to have several optimized implementations
of library routines depending on which hardware extensions are enabled
and the current mechanism to choose the appropriate implementation
will quickly become messy when dealing with many implementations.

With this change, the idea is 'newlib/libc/machine/riscv/memmove.c'
will override the generic 'newlib/libc/string/memmove.c' through the
'srcs_machine' scheme and conditions in 'newlib/libc/machine/riscv/memmove.c'
will decide which implementation gets compiled and used for a given scenario.

I think this mechanism is simple and scalable.