sys/types.h: include <sys/_pthreadtypes.h>
It seems that some essential types are not defined when using
```cpp
include <pthread.h>
```

Indeed, when comparing `sys/types.h` with the Newlib variant,
we see that `sys/_pthreadtypes.h` is missing in Picolibc.

Include the missing file in `sys/types.h` to get access to
other essential posix types.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>