math: lgammaf returned inf for values between -0x1p-70 and -0x1p-64
A range check which tried to short-circuit the computation for small
values had an incorrect bounds of 0x1p-70.

The Euler reflection formula for negative values computes
pi/(sin(πx)*x) as an intermediate term, if |x| was less than 0x1p-64,
the value of sin(πx)*x would be sub-normal, and dividing that into pi
would result in overflow to ∞. Adjust the check to use 0x1p-35, which
still makes the function return values within 2ulp.

Signed-off-by: Keith Packard <keithp@keithp.com>