string: Cast memchr and memrchr val to 'unsigned long' while building mask
In the speed-optimized version of this code, the needle is replicated
to fill an unsigned long. It does that by oring various shifts of the
value. Of course, C performs the first of those by first casting to
int, and on MSP430, the compiler ends up sign extending so values >=
0x80 end up with high bits set which wrecks the resulting mask value.

Signed-off-by: Keith Packard <keithp@keithp.com>