stdlib: Use an alias for malloc internally
When malloc is called from calloc, UBSAN uses the malloc attributes to
compute the bounds of the returned pointer. When calloc attempts to
clear the padding after the allocation, that triggers an error.

Avoid this by using a strong alias of malloc that doesn't include the
attributes. Change the existing free alias to work the same way, in case
the compiler decides to do something similar with that.

Signed-off-by: Keith Packard <keithp@keithp.com>